pyproject.toml
[project]
name = "my-mcp-server"
version = "0.1.0"
description = "FastMCP server with Google OAuth and persistent storage"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp==2.14.1",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
